Created an dropdown in the navbar containing upload status instead of having it on the upload page
