+++ app/Livewire/Upload.php
@@ -32,12 +32,6 @@ class Upload extends Component
     #[Validate('image')]
     public $image;
 
-    #[Locked]
-    public $completedImageUploads = [];
-
-    #[Locked]
-    public $imageUploads = [];
-
     public function onUploadFinished() {
         $img = ImageManager::gd()->read($this->image);
 
@@ -59,38 +53,6 @@ class Upload extends Component
         $this->image = null;
 
         return $this->redirectRoute('upload.process', ['uuid' => $upload->uuid], navigate: true);
-    }
-
-    public function retrieveImageUploads()
-    {
-        $uploads = ImageUpload::where('user_id', Auth::user()->id)->orderBy('created_at', 'desc')->get();
-        $this->imageUploads = [];
-        $this->completedImageUploads = [];
-        foreach ($uploads as $upload)
-        {
-            if($upload->state === "done")
-                $this->completedImageUploads[$upload->uuid] = ['state' => $upload->state, 'startTime' => $upload->created_at->diffForHumans()];
-            else
-                $this->imageUploads[$upload->uuid] = ['state' => $upload->state, 'startTime' => $upload->created_at->diffForHumans()];
-        }
-    }
-
-    public function mount(){
-        $this->retrieveImageUploads();
-    }
-
-    public function getStateColour($state)
-    {
-        switch($state){
-            case 'waiting':
-                return 'border-gray-600';
-            case 'scanning':
-            case 'processing':
-                return 'border-orange-500';
-            case 'foundDuplicates':
-            case 'error':
-                return 'border-red-600';
-        }
     }
 
     public function render()
